Switch to long

+++ src/Main.java
@@ -1,8 +1,8 @@
 public class Main {
     public static void main(String[] args) {
-        int[] sales = new int[] {6, 4, 3};
+        long[] sales = new long[] {6, 4, 3};
         SalesManager salesManager2 = new SalesManager(sales);
-        int max = salesManager2.max();
+        long max = salesManager2.max();
         System.out.println(max);
     }
 }
